Fix the build
aa79cf8743b6815cbc3a57a2f813564fb7f4da3c appeared to cleanly rebase,
becoming 0b0aa9de4e6dee82c7cbfea1145a730c019d1573. But this was a lie!
The guts of stimes became more monomorphic in 0.12, so this commit
removed a 'fromIntegral' that is necessary only in the 0.11 branch.

Not catching this before pushing was definitely my bad. Sorry!

+++ Data/ByteString/Internal/Type.hs
@@ -857,7 +857,7 @@ times n (BS fp len)
   | len == 1 = unsafeCreateFp size $ \destfptr -> do
       byte <- peekFp fp
       unsafeWithForeignPtr destfptr $ \destptr ->
-        fillBytes destptr byte n
+        fillBytes destptr byte (fromIntegral n)
   | otherwise = unsafeCreateFp size $ \destptr -> do
       memcpyFp destptr fp len
       fillFrom destptr len
